+++ src/main/java/com/example/onlyone/domain/notification/dto/responseDto/NotificationCreateResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.onlyone.domain.notification.dto;
+package com.example.onlyone.domain.notification.dto.responseDto;
 
 import com.example.onlyone.domain.notification.entity.Notification;
 import lombok.Builder;
