+++ src/main/java/com/example/onlyone/domain/notification/dto/responseDto/NotificationListResponseDto.java
@@ -1,5 +1,6 @@
-package com.example.onlyone.domain.notification.dto;
+package com.example.onlyone.domain.notification.dto.responseDto;
 
+import com.example.onlyone.domain.notification.dto.requestDto.NotificationItemDto;
 import lombok.Builder;
 import lombok.Getter;
